refactor(lang): Split off object.py
